Adiciona métodos deleteCharge e getCharge ao contrato PaymentGateway

+++ src/slices/payment/charge/useCases/deleteCharge/DeleteCharge.ts
@@ -1,10 +1,17 @@
 import { DeleteChargeRepository } from "@/slices/payment/charge/repositories";
 import { ChargeData } from "@/slices/payment/charge/entities";
 import { Query } from "@/application/types";
+import { PaymentGateway } from "@/application/infra/contracts";
 
 export type DeleteCharge = (query: Query) => Promise<ChargeData | null>;
-export type DeleteChargeSignature = (deleteCharge: DeleteChargeRepository) => DeleteCharge;
+export type DeleteChargeSignature = (
+  deleteCharge: DeleteChargeRepository,
+  paymentProvider: PaymentGateway
+) => DeleteCharge;
 export const deleteCharge: DeleteChargeSignature =
-  (deleteChargeRepository: DeleteChargeRepository) => (query: Query) => {
+  (deleteChargeRepository: DeleteChargeRepository, paymentProvider: PaymentGateway) =>
+  async (query: Query) => {
+    const response = await paymentProvider.deleteCharge(query?.fields?._id);
+    if (response?.status !== "OK") return null;
     return deleteChargeRepository.deleteCharge(query);
   };
